Setup database connections

There is a src_*() function to setup a connection to each of the database backends. The backends may have specific parameters in the respective function src_*(), but all other nodbi functions are independent of the backend (e.g., see docdb_query()).

Details

MongoDB - src_mongo()

SQLite - src_sqlite()

Elasticsearch - src_elastic()

CouchDB - src_couchdb()

PostgreSQL - src_postgres()

DuckDB - src_duckdb()

Documentation details for each database:

MongoDB - https://docs.mongodb.com/

SQLite/JSON1 - https://www.sqlite.org/json1.html

Elasticsearch - https://www.elastic.co/guide/en/elasticsearch/reference/current/index.html

CouchDB - http://docs.couchdb.org/

PostgreSQL - https://www.postgresql.org/docs/current/functions-json.html

DuckDB - https://duckdb.org/docs/extensions/json

Documentation of R packages used by nodbi for the databases:

mongolite - https://CRAN.R-project.org/package=mongolite

RSQLite - https://CRAN.R-project.org/package=RSQLite

elastic - https://CRAN.R-project.org/package=elastic

sofa - https://CRAN.R-project.org/package=sofa

RPostgres - https://CRAN.R-project.org/package=RPostgres

duckdb - https://CRAN.R-project.org/package=duckdb